requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
sqlalchemy==2.0.23
pydantic==2.5.2
pydantic-settings==2.1.0
python-dotenv==1.0.0
alembic==1.12.1
psycopg2-binary==2.9.9
cryptography==41.0.5
pyotp==2.9.0
qrcode==7.4.2
pillow==10.1.0
python-ldap==3.4.3
ldap3==2.9.1 